Crate conrod_gfx

A gfx backend for rendering conrod primitives.

Structs

Commands
An iterator yielding Commands, produced by the Renderer::commands method.

Renderer
This type is used for translating render::Primitives into Commandss that indicate how to draw the GUI using gfx.

Enums

Command
A Command describing a step in the drawing process.

Draw
A Command for drawing to the target.

RendererCreationError
Possible errors that may occur during a call to Renderer::new.

Constants

MODE_GEOMETRY
Ignore tex and draw simple, colored 2D geometry.

MODE_IMAGE
Draw an image from the texture at tex in the fragment shader.

MODE_TEXT
Draw text from the text cache texture tex in the fragment shader.

Type Definitions

ColorFormat
Color format used with gfx buffers.